GPLIM-1819 concentration is persisted as a Float so why have webservice return double. add FCT Name to fetchSequencingTemplate WS when available

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/boundary/lims/LimsQueryObjectFactory.java
@@ -63,7 +63,7 @@ public class LimsQueryObjectFactory {
      * @param startingVesselLabel
      */
     public static SequencingTemplateLaneType createSequencingTemplateLaneType(@Nonnull String laneName,
-                                                                              @Nullable Double loadingConcentration,
+                                                                              @Nullable Float loadingConcentration,
                                                                               @Nonnull String loadingVesselLabel,
                                                                               String startingVesselLabel) {
         SequencingTemplateLaneType lane = new SequencingTemplateLaneType();
